requirements.txt
pyyaml
openai
python-dotenv
requests
sqlalchemy
psycopg[binary]
beautifulsoup4
